% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampkg.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Building map of incidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number}

\item{year}{The year for data collecting}
}
\value{
make a map in window
}
\description{
This function draws a map with incidents
for the selected state and year
}
\examples{
\dontrun{fars_map_state(4, 2015)}

}
