% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampkg.R
\name{fars_read}
\alias{fars_read}
\title{File reading function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that contain file name}
}
\value{
This function returns a "tbl_df" object
}
\description{
This is a function that reading files and returne "tbl_df" object
}
\examples{
\dontrun{fars_read(filename = "accident_2015.csv.bz2")}

}
