% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampkg.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Return summaries output by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{summarized vector of the years}
}
\value{
"tbl_df" with summary data showing the
number of incidents by month and year.
}
\description{
Return summaries output by years
}
\examples{
\dontrun{fars_summarize_years(2015)}

}
