% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampkg.R
\name{make_filename}
\alias{make_filename}
\title{File name making function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer with year}
}
\value{
This function returns character with accident file name
}
\description{
This is a function that reading files and returns string with complete
file name object
}
\examples{
\dontrun{make_filename(2014)}

}
